requirements.txt
# Core
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.7.4
python-dotenv>=1.2.1

# LangChain
langchain>=1.2.6
langchain-core>=1.2.7
langchain-aws>=0.2.0
langchain-postgres>=0.0.16
langchain-text-splitters>=1.1.0
langgraph>=1.0.6

# AWS
boto3>=1.35.0
botocore>=1.35.0

# Database
psycopg2-binary>=2.9.0
asyncpg>=0.30.0

# PDF Processing
pypdf>=6.0.0

# Utilities
pgvector>=0.3.0